migrate: move maintenanceBanner to components directory
- Migrated maintenanceBanner component files to src/components/maintenanceBanner/
- Moved and renamed index.tsx -> template.tsx
- Moved and renamed index.test.tsx -> template.test.tsx
- Updated import statements:
  - src/components/header/template.tsx from @/templates/common/banners/maintenanceBanner to @/components/maintenanceBanner/template
  - template.test.tsx: ./ to ./template
- Removed empty src/templates/common/banners/maintenanceBanner/ directory
- All TypeScript compilation checks passing
- All tests passing (1 test for MaintenanceBanner component)

This maintenanceBanner component is a simple widget container for maintenance banners
used in the header layout. It renders a div with maintenance-banner data attributes.
